Correct prototypes for POSIX.
git-svn-id: https://svn.aros.org/svn/aros/trunk@20685 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/clib/lseek.c
@@ -15,7 +15,7 @@
     NAME */
 #include <unistd.h>
 
-	int lseek (
+	off_t lseek (
 
 /*  SYNOPSIS */
 	int    filedes,
